RomanNumerals/RomanNumeralGenerator.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace RomanNumerals
{
    public class RomanNumeralGenerator
    {
        public readonly Dictionary<int, string> NumberRomanDictionary;

        public RomanNumeralGenerator()
        {
            NumberRomanDictionary = new Dictionary<int, string>
            {
                { 1000, "M" },
                { 900, "CM" },
                { 500, "D" },
                { 400, "CD" },
                { 100, "C" },
                { 50, "L" },
                { 40, "XL" },
                { 10, "X" },
                { 9, "IX" },
                { 5, "V" },
                { 4, "IV" },
                { 1, "I" },
            };
        }
        public string GetRomanNumeral(int number)
        {
            if (number <= 0 || number > 50)
            {
                throw new Exception("Must provide a positive number between 1 and 50");
            }
            var roman = new StringBuilder();

            foreach (var item in NumberRomanDictionary)
            {
                while (number >= item.Key)
                {
                    roman.Append(item.Value);
                    number -= item.Key;
                }
            }

            return roman.ToString();
        }
    }
}
